refactor: update UpdateTaskUseCase to use direct calls & the related imports
